requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
attrs==25.3.0
click==8.2.1
cloudpickle==3.1.1
dask==2025.5.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lightning-utilities==0.14.3
locket==1.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
numpy
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
partd==1.4.2
pillow==11.3.0
propcache==0.3.2
pycocotools==2.0.10
python-dotenv==1.1.1
pytorch-lightning==2.5.2
PyYAML==6.0.2
scikit-learn==1.7.0
scipy
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
transformers
tensorboard
toolz==1.0.0
torch==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
tqdm==4.67.1
trimesh==4.6.13
triton==3.3.1
typing_extensions==4.14.1
yarl==1.20.1
